refactor(texts): file service
