--- src/modules/texts/Text/service/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./Text.service";
-export * from "./Text.service";
