+++ src/modules/texts/Text/services/index.ts
@@ -1,1 +1,2 @@
 export * from "./fetchGetAllTextsService";
+export * from "./fetchPostTextService";
