+++ src/modules/texts/Text/services/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./fetchGetAllTextsService.types";
+export * from "./fetchPostTextService.types";
